add second reviewer, date filter and use cursor to fetch all search results
I modified the original script to use the cursor, which allows the
search results to be pulled in smaller chunks. I found that Github will
often time out if fetching 100 results at a time but using the cursor
with smaller chunks made life a bit nicer.

In addition I wanted to see the second reviewer and be able to filter
by date so hacked them in too.

All a bit janky but seems to work. Decided if I want any deeper
analysis I'd probably change the approach as Github GraphQL has too
many limitations.